Backend/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-doc
annotated-types
anyio
attrs
bcrypt
cachetools
certifi
charset-normalizer
click
colorama
dataclasses-json
dnspython
email-validator
fastapi
filelock
filetype
frozenlist
fsspec
google-ai-generativelanguage
google-api-core
google-auth
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
httpcore
httpx
httpx-sse
huggingface-hub
idna
Jinja2
joblib
jsonpatch
jsonpointer
langchain
langchain-classic
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
langchain-ollama
langchain-pinecone
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
lxml
MarkupSafe
marshmallow
mpmath
multidict
mypy_extensions
networkx
numpy
ollama
orjson
ormsgpack
packaging
passlib
pdf2image
pillow
pinecone
pinecone-plugin-assistant
pinecone-plugin-interface
propcache
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic-settings
pydantic_core
PyJWT
PyPDF2
pytesseract
python-dateutil
python-docx
python-dotenv
python-multipart
PyYAML
regex
requests
requests-toolbelt
rsa
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
simsimd
six
sniffio
starlette
sympy
tenacity
threadpoolctl
tokenizers
torch
tqdm
transformers
typing-inspect
typing-inspection
typing_extensions
urllib3
uvicorn
xxhash
yarl
zstandard
SQLAlchemy
